% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui-widget.R
\name{liberality_workbench}
\alias{liberality_workbench}
\title{LibeRality React workbench widget}
\usage{
liberality_workbench(
  payload,
  input_id = "liberality_workbench",
  width = NULL,
  height = "100vh",
  elementId = NULL
)
}
\arguments{
\item{payload}{Workbench payload.}

\item{input_id}{Shiny event prefix.}

\item{width, height}{Widget dimensions.}

\item{elementId}{Optional element id.}
}
\description{
LibeRality React workbench widget
}
