% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui-widget.R
\name{liberalityWorkbenchOutput}
\alias{liberalityWorkbenchOutput}
\title{Shiny output for the LibeRality workbench}
\usage{
liberalityWorkbenchOutput(outputId, width = "100\%", height = "100vh")
}
\arguments{
\item{outputId}{Output id.}

\item{width, height}{CSS dimensions.}
}
\description{
Shiny output for the LibeRality workbench
}
