% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables-constraints.R
\name{lity_constraint}
\alias{lity_constraint}
\title{Define a design constraint}
\usage{
lity_constraint(
  name,
  type = c("min_separation", "max_samples", "total_subjects", "total_cost",
    "max_blood_volume", "max_duration", "arm_size", "allocation", "max_rse",
    "minimum_power", "exposure", "custom"),
  limit,
  arm = NULL,
  endpoint = NULL,
  parameters = NULL,
  lower = NULL,
  upper = NULL,
  function_value = NULL,
  metadata = list()
)
}
\arguments{
\item{name}{Constraint name.}

\item{type}{Constraint type.}

\item{limit}{Upper or lower limit, depending on `type`.}

\item{arm}{Optional arm.}

\item{endpoint}{Optional endpoint.}

\item{parameters}{Optional parameter selection for precision constraints.}

\item{lower, upper}{Optional interval bounds.}

\item{function_value}{Custom function accepting a design and returning a scalar.}

\item{metadata}{Additional metadata.}
}
\description{
Define a design constraint
}
