% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables-constraints.R
\name{lity_constraint_check}
\alias{lity_constraint_check}
\title{Evaluate all design constraints}
\usage{
lity_constraint_check(design, evaluation = NULL)
}
\arguments{
\item{design}{Design object.}

\item{evaluation}{Optional cached design evaluation.}
}
\description{
Evaluate all design constraints
}
