% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{lity_endpoint_from_liberator}
\alias{lity_endpoint_from_liberator}
\title{Convert a LibeRator endpoint into a LibeRality design endpoint}
\usage{
lity_endpoint_from_liberator(endpoint, dvid = 1L, type = "continuous")
}
\arguments{
\item{endpoint}{A `lator_endpoint`.}

\item{dvid}{Outcome identifier.}

\item{type}{Statistical response type used by the design model.}
}
\description{
Convert a LibeRator endpoint into a LibeRality design endpoint
}
