% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{lity_evaluate}
\alias{lity_evaluate}
\title{Evaluate one or more criteria for a design}
\usage{
lity_evaluate(
  design,
  criteria = lity_criterion_D(),
  tolerance = 1e-10,
  check_constraints = TRUE
)
}
\arguments{
\item{design}{Design object.}

\item{criteria}{Criterion or named list of criteria.}

\item{tolerance}{Numerical rank tolerance.}

\item{check_constraints}{Evaluate constraints after criteria.}
}
\value{
A `lity_evaluation`.
}
\description{
Evaluate one or more criteria for a design
}
