% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{lity_model_from_liberary}
\alias{lity_model_from_liberary}
\title{Import a reviewed LibeRary model}
\usage{
lity_model_from_liberary(library_id, root = NULL, allow_draft = FALSE)
}
\arguments{
\item{library_id}{Catalogue identifier.}

\item{root}{Optional catalogue root.}

\item{allow_draft}{Permit a draft entry.}
}
\description{
Import a reviewed LibeRary model
}
