% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{lity_operating_characteristics}
\alias{lity_operating_characteristics}
\title{Summarise empirical operating characteristics}
\usage{
lity_operating_characteristics(simulation, alpha = 0.05)
}
\arguments{
\item{simulation}{A result from [lity_simulate_trials()].}

\item{alpha}{Confidence interval alpha.}
}
\description{
Summarise empirical operating characteristics
}
