% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{lity_population}
\alias{lity_population}
\title{Define populations and covariate strata}
\usage{
lity_population(
  strata = data.frame(name = "default", weight = 1),
  name = "Population",
  metadata = list()
)
}
\arguments{
\item{strata}{Data frame with a stratum `name`, `weight`, and optional model covariates.}

\item{name}{Population set name.}

\item{metadata}{Additional metadata.}
}
\description{
Define populations and covariate strata
}
