% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{lity_validate}
\alias{lity_validate}
\title{Validate a LibeRality design}
\usage{
lity_validate(design, strict = FALSE)
}
\arguments{
\item{design}{Design object.}

\item{strict}{Whether warnings should be returned as validation failures.}
}
\value{
Validation report.
}
\description{
Validate a LibeRality design
}
