% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{lity_worker_task}
\alias{lity_worker_task}
\title{Execute a typed LibeRality worker task}
\usage{
lity_worker_task(model, design, arguments = list())
}
\arguments{
\item{model}{Serialized primary model retained by the queue contract.}

\item{design}{Serialized design.}

\item{arguments}{Worker arguments.}
}
\description{
Execute a typed LibeRality worker task
}
